third_party/blink/web_tests/platform/mac/virtual/backface-visibility-interop/external/wpt/css/css-transforms/parsing/webkit-perspective-invalid.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['-webkit-perspective'] = "calc(1000)" should not set the property value
  assert_equals: expected "" but got "1000px"
[FAIL] e.style['-webkit-perspective'] = "calc(25)" should not set the property value
  assert_equals: expected "" but got "25px"
Harness: the test ran to completion.
